Changed: merge session module into run_hook
- Moved the run-lifecycle types (`EndReason`, `HookRunContext`, `SessionCompactFn`) from the removed `hooks/session/` module into `hooks/run_hook/`, so the run hook chain owns its context types like `tool_hook` does.
- Crate-root public API is unchanged; this is a pure relocation.
- Folded the `hooks/mod.rs` doc list.
- Fixed two test-only imports that referenced the removed module.
